[api/v2/test] minor fix & it is now tidier [#325]
