+++ ethicapp/backend/api/v2/testApi.js
@@ -5,7 +5,7 @@ const designRouter = require('./designs')
 const userRouter = require('./users')
 const sessionRouter = require('./sessions')
 const loginUserRouter = require('./login-user')
-const groupRouter = require('./groups')
+const groupRouter = require('./teams')
 const chatroomRouter = require('./chatrooms')
 const activityRouter = require('./activities')
 const sessionUserRouter = require('./sessions-users')
